Changed optimization from -Os to -O1 due to problems with memory garbage, optimized and refactored some parts of code from perspective of flash size, functional settings screen added(for now only pwm can be changed, but in the future current, temperature and temperature can be also added)

+++ Core/Inc/ftoa.h
@@ -16,6 +16,6 @@
  * @param res Pointer to the output character array.
  * @param afterpoint The number of digits to display after the decimal point.
  */
-void ftoa(float n, char* res, int afterpoint);
+void ftoa(float n, char* res, uint8_t afterpoint);
 
 #endif /* INC_FTOA_H_ */
